drm: ensure modesetting driver supports 32bpp fb
Now that more and more hardware is getting kms support,
we need to be more careful about what hardware we try to
run on.  Cirrus cards can't make 32bpp framebuffers, which
is all we support in the drm renderer.

This commit tries to create a small buffer up front, just
to see if the driver will allow it.  If the driver chokes,
then we bail and fall back to the /dev/fb renderer, which
is more tolerant of aging hardware.