drm: use queried minimum size, not 1x1 for dummy buffer
The previous commit tries to create a 32bpp 1x1 dummy buffer
to know whether or not the modesetting driver can handle
our needs.  It's concievable (though not realized in practice) that
the driver could support our needs fine, but not support buffers
as small as 1x1.

This commit uses the queried minimum buffer dimensions for the dummy
buffer instead of assuming 1x1.