main: drop parameters for check_for_consoles function
both the default tty and whether or not displays should be added
are available in the global state object (as state->default_tty
and state->is_shown respectively), so they don't need to be parameters.

The latter one was a boolean, so having it as a parameter was confusing
at the call sites.

This commit drops the parameters.