renderer: use enum for selecting renderer type, not plugin path
ply_renderer_new takes a path to a renderer plugin, or NULL to
try each one in turn.

It's cleaner to abstract the path behind an enum type, so this
commit makes that change.

Now it will be possible to instantiate specific renderers without
hardcoding the paths to plugins in more than one place.