renderer: add way to get device_name that was passed to constructor
It's useful to be able to figure out which renderer a given renderer
is, by examining the device that is associated with it.

This commit adds and accessor function to return the device that
was passed to ply_renderer_new.

At the moment, it does not return the device name if NULL was passed
to the constructor and the device was figured out automatically.  A
future commit may add that ability if it becomes necessary.