build-goo: link libudev with libply-splash-core instead of plymouthd
The libply-splash-core library provides functions for controlling the
input and output of the splash screen.

Recently it gained support for device enumeration from udev using the
libudev library. When that support was added, the linker and compiler
flags of the plymouthd binary were augmented, instead of the the
libply-splash-core library directly. That broke the build on some
systems.

This commit moves the linker and compiler flags to the correct place.