animation: don't report success when no frames were added
Currently, the animation object will report success when assets are *found* in
the directory. When loading those assets, code will apply more checks to
determine if the assets are meant for the animation.
In case none of them are, we'll end up not adding any frames, so we
should not report a successful load.

This is like commit 741b545868aa5bd42a64d5316c121fe9bae4bc2f but for
animation objects instead of throbber objects.