ply-pixel-buffer: Add ply_pixel_buffer_rotate_upright helper
For some themes we want to use the firmware-logo / splash as background,
when the LCD panel of a laptop is mounted non-upright then the image which
we get from the firmware will be pre-rotated to match the LCD panel mount.

Until now our device-rotation support was limited to using rotated
pixel_buffer-s as destination / canvas only.

This commit adds a ply_pixel_buffer_rotate_upright helper to rotate
a nop-upright source buffer upright so that we can use it as source-buffer
to other functions which expect source-buffers to always be upright.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>